refactor(tests): split WPLoaderTest and SingleTest to rebalance parallel shards
WPLoaderTest.php (2931 lines, 169s) split into 5 focused files
(ConfigValidation, Paths, Bootstrap, DbDumpsSqliteBedrock, ThemeBackup) via
a shared WPLoaderSetup trait holding $config/$previousCwd/HOME backup
lifecycle and the module() helper. SingleTest.php (1143 lines, 113s) split
into 4 (Construction, MultisiteConversion, Directories, Runtime).

Also narrows the file_get_contents mock in
should_throw_if_wp_config_php_file_contents_cannot_be_read_during_multsite_conversion
to only the wp-config.php path. The blanket mock was poisoning
Opis ReflectionClosure's static $files token cache for Single.php,
making every later ReflectionClosure->isStatic() call in the process
return false. That pollution was previously masked by earlier tests in
the monolithic SingleTest priming the cache first; splitting exposed it.

After the split, max shard weight drops from 169s to ~35s and the
parallel-run unit suite goes from 130s (41 failures) to ~94s (0 failures).
